feat(api): migrate controller route campaign overviews
