+++ api/src/prescription/campaign-participation/infrastructure/serializers/jsonapi/campaign-participation-overview-serializer.js
