+++ api/tests/unit/application/users/user-controller_test.js
@@ -170,85 +170,6 @@ describe('Unit | Controller | user-controller', function () {
     });
   });
 
-  describe('#getCampaignParticipationOverviews', function () {
-    const userId = '1';
-    let dependencies;
-
-    beforeEach(function () {
-      const campaignParticipationOverviewSerializer = {
-        serialize: sinon.stub(),
-        serializeForPaginatedList: sinon.stub(),
-      };
-      sinon.stub(usecases, 'findUserCampaignParticipationOverviews');
-      dependencies = {
-        campaignParticipationOverviewSerializer,
-      };
-    });
-
-    it('should return serialized campaignParticipationOverviews', async function () {
-      // given
-      const request = {
-        auth: {
-          credentials: {
-            userId: userId,
-          },
-        },
-        params: {
-          id: userId,
-        },
-        query: { filter: {}, page: {} },
-      };
-      usecases.findUserCampaignParticipationOverviews.withArgs({ userId, states: undefined, page: {} }).resolves([]);
-      dependencies.campaignParticipationOverviewSerializer.serializeForPaginatedList.withArgs([]).returns({
-        id: 'campaignParticipationOverviews',
-      });
-
-      // when
-      const response = await userController.getCampaignParticipationOverviews(request, hFake, dependencies);
-
-      // then
-      expect(response).to.deep.equal({
-        id: 'campaignParticipationOverviews',
-      });
-      expect(dependencies.campaignParticipationOverviewSerializer.serializeForPaginatedList).to.have.been.calledOnce;
-    });
-
-    it('should forward state and page query parameters', async function () {
-      // given
-      const request = {
-        auth: {
-          credentials: {
-            userId: userId,
-          },
-        },
-        params: {
-          id: userId,
-        },
-        query: {
-          filter: {
-            states: 'ONGOING',
-          },
-          page: { number: 1, size: 10 },
-        },
-      };
-      usecases.findUserCampaignParticipationOverviews
-        .withArgs({ userId, states: 'ONGOING', page: { number: 1, size: 10 } })
-        .resolves([]);
-      dependencies.campaignParticipationOverviewSerializer.serializeForPaginatedList.withArgs([]).returns({
-        id: 'campaignParticipationOverviews',
-      });
-
-      // when
-      const response = await userController.getCampaignParticipationOverviews(request, hFake, dependencies);
-
-      // then
-      expect(response).to.deep.equal({
-        id: 'campaignParticipationOverviews',
-      });
-      expect(dependencies.campaignParticipationOverviewSerializer.serializeForPaginatedList).to.have.been.calledOnce;
-    });
-  });
-
   describe('#resetScorecard', function () {
     beforeEach(function () {
       sinon.stub(evaluationUsecases, 'resetScorecard').resolves({
